+++ .gitignore
@@ -1,2 +1,3 @@
 preprocessing\normal_features.csv
 preprocessing\emails-normal-class.csv
+dataset\phishing_email.csv
